relocate stake helper to avoid cycles
