+++ libs/model/src/utils/index.ts
@@ -2,6 +2,5 @@ export * from './denormalizedCountUtils';
 export * from './getDelta';
 export * from './parseUserMentions';
 export * from './sanitizeQuillText';
-export * from './stakeHelper';
 export * from './updateTags';
 export * from './utils';
